+++ terra-theme.config.js
@@ -1,5 +1,5 @@
 const themeConfig = {
-  scoped: ['orion-fusion-theme'],
+  scoped: ['clinical-lowlight-theme', 'orion-fusion-theme'],
 };
 
 module.exports = themeConfig;