+++ packages/terra-progress-bar/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added clinical-lowlight-theme theming files
 
 4.18.0 - (January 7, 2020)
 ------------------