+++ packages/terra-progress-bar/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+4.19.0 - (February 18, 2020)
+------------------
+### Added
+* Added clinical-lowlight-theme theming files
+
 4.18.0 - (January 7, 2020)
 ------------------
 ### Removed